added logout and fix bugs with routes

+++ src/router.js
@@ -1,14 +1,16 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
+// import requests from './requests';
 
 import { store } from './store/index';
-import axios from 'axios';
+// import axios from 'axios';
+import guard from './requests'
 
 Vue.use(VueRouter);
 
-const guard = axios.create({
-    baseURL: 'https://xeniaweb.online/api',    
-});
+// const guard = axios.create({
+//     baseURL: 'https://xeniaweb.online/api',    
+// });
 const userId = localStorage.getItem('userId');
 
 const routes = [
@@ -44,32 +46,28 @@ const router = new VueRouter({
 });
 
 router.beforeEach(async (to, from, next) => {
-    // console.log('tut');
-    // store.actions['user/login()'];
     const isPublicRoute = to.matched.some(route => route.meta.public);
+    const isUserLogged = store.getters['user/userIsLogged'];
+    console.log('isUserLogged',isUserLogged);
 
-    // console.log('tut isUserLogged', isUserLogged);
-    // const isUserLogged = store.getters['user/userIsLogged'];
-
-
-    if (!isPublicRoute) {
-    // if (!isPublicRoute && !isUserLogged) {
+    if (!isPublicRoute && !isUserLogged) {
         const token = localStorage.getItem('token');
         guard.defaults.headers["Authorization"] = `Bearer ${token}`;
 
        try {
             const {data} = await guard.get('/user');
-            console.log('Ответ от сервера: ',data);
-            store.commit('user/SET_USER', data.user);
+            console.log('servers answer: ', data);
+            store.commit('user/SET_USER', data);
+            // userId = data.id;
+            localStorage.setItem('userId', data.id);
             next();
-       } 
-       catch (error) {        
+        } 
+        catch (error) {        
             router.replace('/login');
             localStorage.clear();
-       }
+        }
     }
     else {
-        console.log('next()');
         next();
     }
